update to ipldv7
* use newer graphql codegen
* use newer ipld

+++ types/ipldsch_minima.go
@@ -35,11 +35,11 @@ type _ErrorThunkAssembler struct {
 	e error
 }
 
-func (ea _ErrorThunkAssembler) BeginMap(_ int) (ipld.MapAssembler, error) { return nil, ea.e }
-func (ea _ErrorThunkAssembler) BeginList(_ int) (ipld.ListAssembler, error) { return nil, ea.e }
+func (ea _ErrorThunkAssembler) BeginMap(_ int64) (ipld.MapAssembler, error) { return nil, ea.e }
+func (ea _ErrorThunkAssembler) BeginList(_ int64) (ipld.ListAssembler, error) { return nil, ea.e }
 func (ea _ErrorThunkAssembler) AssignNull() error { return ea.e }
 func (ea _ErrorThunkAssembler) AssignBool(bool) error { return ea.e }
-func (ea _ErrorThunkAssembler) AssignInt(int) error { return ea.e }
+func (ea _ErrorThunkAssembler) AssignInt(int64) error { return ea.e }
 func (ea _ErrorThunkAssembler) AssignFloat(float64) error { return ea.e }
 func (ea _ErrorThunkAssembler) AssignString(string) error { return ea.e }
 func (ea _ErrorThunkAssembler) AssignBytes([]byte) error { return ea.e }
